Add files via upload
We will model a real life situation that you are familiar with: registering for courses at
university! We will model a set of courses and a set of students who register for the courses. We will also
model details such as a course waitlist. (Note we will use the term “course’ rather than “class” so that the
latter term can be reserved for talking about the Java classes.)